Changed the control and added a ignore file
