+++ robot_controls/main.py
@@ -1,9 +1,10 @@
 #!/usr/bin/env pybricks-micropython
 import socket
 import threading
+import subprocess
 from commandparser import parse_and_execute
 
-bind_ip = "192.168.28.33"
+bind_ip = "192.168.8.139"
 bind_port = 8080  
 
 # create and bind a new socket
@@ -12,12 +13,14 @@ server.bind((bind_ip, bind_port))
 server.listen(5)
 print("Server is listening on %s:%d" % (bind_ip, bind_port))
 
+
 def clientHandler(client_socket):
     client_socket.send("ready".encode())
     
     request = client_socket.recv(1024)
     print("Received \"" + request.decode() + "\" from client")
     command = request.decode()
+    parse_and_execute(command)
 
 while True:
     # wait for client to connect
